requirements.txt
dbus-python==1.3.2
distro==1.8.0
gpg==1.17.1
ImageHash==4.3.1
libcomps==0.1.18
numpy==1.25.2
Pillow==10.0.1
python-dateutil==2.8.2
PyWavelets==1.4.1
rpm==4.18.1
scipy==1.11.2
six==1.16.0
systemd-python==235
tqdm==4.66.1